+++ src/components/PostInfo/PostInfo.tsx
@@ -20,13 +20,7 @@ export const PostInfo: React.FC<Props> = ({ post }) => {
     <div className="PostInfo">
       <div className="PostInfo__header">
         <h3 className="PostInfo__title">{title}</h3>
-
-        <p>
-          {' Posted by  '}
-
-          {user && <UserInfo user={user} />}
-
-        </p>
+        {user && <UserInfo user={user} />}
       </div>
 
       <p className="PostInfo__body">
