+++ src/components/UserInfo/UserInfo.tsx
@@ -10,8 +10,13 @@ export const UserInfo: React.FC<Props> = ({ user }) => {
   const { name, email } = user;
 
   return (
-    <a className="UserInfo" href={`mailto:${email}`}>
-      {`${name}`}
-    </a>
+    <p>
+      {' Posted by  '}
+
+      <a className="UserInfo" href={`mailto:${email}`}>
+        {`${name}`}
+      </a>
+
+    </p>
   );
 };
